use a shared base test
