+++ 5calls/app/src/androidTest/java/org/a5calls/android/a5calls/controller/MainActivityBaseTest.java
@@ -1,0 +1,116 @@
+package org.a5calls.android.a5calls.controller;
+
+import androidx.test.core.app.ActivityScenario;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
+import androidx.test.platform.app.InstrumentationRegistry;
+
+import com.android.volley.RequestQueue;
+import com.android.volley.toolbox.BasicNetwork;
+
+import org.a5calls.android.a5calls.AppSingleton;
+import org.a5calls.android.a5calls.model.AccountManager;
+import org.a5calls.android.a5calls.net.FakeRequestQueue;
+import org.a5calls.android.a5calls.net.FiveCallsApi;
+import org.a5calls.android.a5calls.net.MockHttpStack;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.runner.RunWith;
+
+/**
+ * Base class for MainActivity integration tests that contains shared setup and utility methods.
+ */
+@RunWith(AndroidJUnit4.class)
+public abstract class MainActivityBaseTest {
+
+    protected MockHttpStack mHttpStack;
+    protected RequestQueue mOriginalRequestQueue;
+    protected FiveCallsApi mOriginalApi;
+    protected String mOriginalAddress;
+    protected ActivityScenario<MainActivity> scenario;
+
+    @Before
+    public void setUp() {
+        // Save original state
+        mOriginalRequestQueue = AppSingleton.getInstance(
+                InstrumentationRegistry.getInstrumentation().getTargetContext()).getRequestQueue();
+        mOriginalApi = AppSingleton.getInstance(
+                InstrumentationRegistry.getInstrumentation().getTargetContext()).getJsonController();
+
+        // Save original location
+        mOriginalAddress = AccountManager.Instance.getAddress(
+                InstrumentationRegistry.getInstrumentation().getTargetContext());
+
+        // Set a mock location to avoid location prompts
+        AccountManager.Instance.setAddress(
+                InstrumentationRegistry.getInstrumentation().getTargetContext(),
+                "90210");
+
+        // Mark tutorial as seen to bypass onboarding screen
+        AccountManager.Instance.setTutorialSeen(
+                InstrumentationRegistry.getInstrumentation().getTargetContext(),
+                true);
+
+        // Create mock HTTP stack
+        mHttpStack = new MockHttpStack();
+    }
+
+    @After
+    public void tearDown() {
+        // Restore original state
+        AppSingleton.getInstance(
+                InstrumentationRegistry.getInstrumentation().getTargetContext())
+                .setRequestQueue(mOriginalRequestQueue);
+        AppSingleton.getInstance(
+                InstrumentationRegistry.getInstrumentation().getTargetContext())
+                .setFiveCallsApi(mOriginalApi);
+
+        // Restore original location
+        AccountManager.Instance.setAddress(
+                InstrumentationRegistry.getInstrumentation().getTargetContext(),
+                mOriginalAddress);
+
+        // Close the activity scenario if it's open
+        if (scenario != null) {
+            scenario.close();
+        }
+    }
+
+    /**
+     * Sets up the mock request queue and API
+     */
+    protected void setupMockRequestQueue() {
+        // Create a custom RequestQueue with our mock HTTP stack
+        BasicNetwork basicNetwork = new BasicNetwork(mHttpStack);
+        FakeRequestQueue requestQueue = new FakeRequestQueue(basicNetwork);
+        requestQueue.start();
+
+        // Replace the app's RequestQueue with our mock
+        AppSingleton.getInstance(
+                InstrumentationRegistry.getInstrumentation().getTargetContext())
+                .setRequestQueue(requestQueue);
+
+        // Create a new FiveCallsApi with our mock RequestQueue
+        String callerId = AccountManager.Instance.getCallerID(
+                InstrumentationRegistry.getInstrumentation().getTargetContext());
+        FiveCallsApi api = new FiveCallsApi(callerId, requestQueue);
+        AppSingleton.getInstance(
+                InstrumentationRegistry.getInstrumentation().getTargetContext())
+                .setFiveCallsApi(api);
+    }
+
+    /**
+     * Launches the MainActivity and waits for it to load
+     * @param waitTimeMs time to wait for the activity to load
+     */
+    protected void launchMainActivity(int waitTimeMs) {
+        // Launch the activity
+        scenario = ActivityScenario.launch(MainActivity.class);
+
+        // Wait for processing and UI to update
+        try {
+            Thread.sleep(waitTimeMs);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+    }
+}
