+++ 5calls/app/src/androidTest/java/org/a5calls/android/a5calls/controller/MainActivityHappyPathTest.java
@@ -1,30 +1,20 @@
 package org.a5calls.android.a5calls.controller;
 
 import android.view.View;
-import androidx.test.core.app.ActivityScenario;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.platform.app.InstrumentationRegistry;
 
-import com.android.volley.RequestQueue;
-import com.android.volley.toolbox.BasicNetwork;
 import com.android.volley.toolbox.HttpResponse;
 import com.google.android.material.appbar.CollapsingToolbarLayout;
 
-import org.a5calls.android.a5calls.AppSingleton;
 import org.a5calls.android.a5calls.FakeJSONData;
 import org.a5calls.android.a5calls.R;
-import org.a5calls.android.a5calls.model.AccountManager;
-import org.a5calls.android.a5calls.net.FakeRequestQueue;
-import org.a5calls.android.a5calls.net.FiveCallsApi;
-import org.a5calls.android.a5calls.net.MockHttpStack;
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -43,13 +33,7 @@ import static org.hamcrest.Matchers.containsString;
  * Integration test for MainActivity that tests the happy path.
  */
 @RunWith(AndroidJUnit4.class)
-public class MainActivityHappyPathTest {
-
-    private MockHttpStack mHttpStack;
-    private RequestQueue mOriginalRequestQueue;
-    private FiveCallsApi mOriginalApi;
-    private String mOriginalAddress;
-    private ActivityScenario<MainActivity> scenario;
+public class MainActivityHappyPathTest extends MainActivityBaseTest {
 
     // Custom matcher to check if a CollapsingToolbarLayout's title contains specific text
     public static Matcher<View> withCollapsingToolbarTitle(final Matcher<String> textMatcher) {
@@ -70,53 +54,6 @@ public class MainActivityHappyPathTest {
                 textMatcher.describeTo(description);
             }
         };
-    }
-
-    @Before
-    public void setUp() {
-        // Save original state
-        mOriginalRequestQueue = AppSingleton.getInstance(
-                InstrumentationRegistry.getInstrumentation().getTargetContext()).getRequestQueue();
-        mOriginalApi = AppSingleton.getInstance(
-                InstrumentationRegistry.getInstrumentation().getTargetContext()).getJsonController();
-
-        // Save original location
-        mOriginalAddress = AccountManager.Instance.getAddress(
-                InstrumentationRegistry.getInstrumentation().getTargetContext());
-
-        // Set a mock location to avoid location prompts
-        AccountManager.Instance.setAddress(
-                InstrumentationRegistry.getInstrumentation().getTargetContext(),
-                "90210");
-
-        // Mark tutorial as seen to bypass onboarding screen
-        AccountManager.Instance.setTutorialSeen(
-                InstrumentationRegistry.getInstrumentation().getTargetContext(),
-                true);
-
-        // Create mock HTTP stack
-        mHttpStack = new MockHttpStack();
-    }
-
-    @After
-    public void tearDown() {
-        // Restore original state
-        AppSingleton.getInstance(
-                InstrumentationRegistry.getInstrumentation().getTargetContext())
-                .setRequestQueue(mOriginalRequestQueue);
-        AppSingleton.getInstance(
-                InstrumentationRegistry.getInstrumentation().getTargetContext())
-                .setFiveCallsApi(mOriginalApi);
-
-        // Restore original location
-        AccountManager.Instance.setAddress(
-                InstrumentationRegistry.getInstrumentation().getTargetContext(),
-                mOriginalAddress);
-
-        // Close the activity scenario if it's open
-        if (scenario != null) {
-            scenario.close();
-        }
     }
 
     /**
@@ -143,44 +80,6 @@ public class MainActivityHappyPathTest {
 
         // Set a default response for any other requests
         mHttpStack.setResponseToReturn(new HttpResponse(200, new ArrayList<>(), "{}".getBytes()));
-    }
-
-    /**
-     * Sets up the mock request queue and API
-     */
-    private void setupMockRequestQueue() {
-        // Create a custom RequestQueue with our mock HTTP stack
-        BasicNetwork basicNetwork = new BasicNetwork(mHttpStack);
-        FakeRequestQueue requestQueue = new FakeRequestQueue(basicNetwork);
-        requestQueue.start();
-
-        // Replace the app's RequestQueue with our mock
-        AppSingleton.getInstance(
-                InstrumentationRegistry.getInstrumentation().getTargetContext())
-                .setRequestQueue(requestQueue);
-
-        // Create a new FiveCallsApi with our mock RequestQueue
-        String callerId = AccountManager.Instance.getCallerID(
-                InstrumentationRegistry.getInstrumentation().getTargetContext());
-        FiveCallsApi api = new FiveCallsApi(callerId, requestQueue);
-        AppSingleton.getInstance(
-                InstrumentationRegistry.getInstrumentation().getTargetContext())
-                .setFiveCallsApi(api);
-    }
-
-    /**
-     * Launches the MainActivity and waits for it to load
-     */
-    private void launchMainActivity(int waitTimeMs) {
-        // Launch the activity
-        scenario = ActivityScenario.launch(MainActivity.class);
-
-        // Wait for all requests to complete and UI to update
-        try {
-            Thread.sleep(waitTimeMs);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
     }
 
     @Test
